Npm update and move hero CSS
Npm update and move hero CSS into main.css so we can include it into gatsby-ssr.js
