+++ package.json
@@ -2,7 +2,7 @@
   "name": "dfweb-v2",
   "private": true,
   "description": "Personal portfolio",
-  "version": "0.3.5",
+  "version": "0.3.6",
   "license": "MIT",
   "scripts": {
     "build": "gatsby build",
@@ -19,17 +19,17 @@
     "@fortawesome/fontawesome-svg-core": "^1.2.29",
     "@fortawesome/free-brands-svg-icons": "^5.13.1",
     "@fortawesome/react-fontawesome": "^0.1.11",
-    "gatsby": "^2.23.12",
-    "gatsby-plugin-manifest": "^2.4.14",
-    "gatsby-plugin-netlify": "^2.3.7",
-    "gatsby-plugin-postcss": "^2.3.6",
+    "gatsby": "^2.23.14",
+    "gatsby-plugin-manifest": "^2.4.16",
+    "gatsby-plugin-netlify": "^2.3.9",
+    "gatsby-plugin-postcss": "^2.3.8",
     "gatsby-plugin-prefetch-google-fonts": "^1.4.3",
     "gatsby-plugin-purgecss": "^5.0.0",
-    "gatsby-plugin-react-helmet": "^3.3.6",
+    "gatsby-plugin-react-helmet": "^3.3.8",
     "gatsby-plugin-react-svg": "^3.0.0",
     "gatsby-plugin-robots-txt": "^1.5.1",
-    "gatsby-plugin-sitemap": "^2.4.7",
-    "gatsby-source-graphql": "^2.5.7",
+    "gatsby-plugin-sitemap": "^2.4.9",
+    "gatsby-source-graphql": "^2.5.9",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-helmet": "^6.1.0",
@@ -40,13 +40,13 @@
   },
   "devDependencies": {
     "babel-eslint": "^10.1.0",
-    "babel-preset-gatsby": "^0.4.12",
+    "babel-preset-gatsby": "^0.5.0",
     "cypress": "^4.9.0",
     "eslint": "^7.3.1",
     "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-react": "^7.20.2",
-    "gatsby-cypress": "^0.4.6",
-    "netlify-cli": "^2.54.0",
+    "eslint-plugin-react": "^7.20.3",
+    "gatsby-cypress": "^0.4.8",
+    "netlify-cli": "^2.57.0",
     "netlify-plugin-cypress": "^1.3.11",
     "prettier": "^2.0.5",
     "react-test-renderer": "^16.13.1",
