+++ src/css/main.css
@@ -11,4 +11,10 @@ body {
   box-shadow: 0 0 0 3px rgba(255, 0, 0, 0.5);
 }
 
+#main-hero
+{
+  background-image: url("../assets//blue-hero.jpg");  
+  height: 450px;
+}
+
 @import "tailwindcss/utilities";
